+++ Source Files/memset.c
@@ -2,7 +2,7 @@
 #include "..\Header Files\memset.h"
 
 /*Function Definition Section*/
-int my_memset(char* str, int val, size_t n)
+size_t my_memset(char* str, int val, size_t n)
 {
 	unsigned int Retval = 0;
 	if (NULL == str)
